requirements.txt
Flask==2.3.3
colorama==0.4.6
pysharedoscom==2.0
python-dotenv==1.0.1
